api/requirements.txt
gunicorn==23.0.0
psycopg2-binary==2.9.10
requests==2.32.3
SQLAlchemy==1.4.54
uvicorn==0.34.0
fastapi==0.110.0
pydantic==2.6.3
pydantic-settings==2.2.1
python-dotenv==1.0.1
slowapi==0.1.9
python-multipart==0.0.9
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
